refactor: split registrar_wizard.py (3538 LOC → 6 modules)
Split god file into package:
- _helpers.py (625 LOC) — imports, schemas, helper functions, router
- _invoice.py (220 LOC) — Click invoice init + status endpoints
- _cart.py (822 LOC) — cart create, edit-delta, repeat-eligibility,
  all-free approval, price-override endpoints
- _settings.py (289 LOC) — benefit/wizard settings endpoints,
  VisitResponse schema
- _visits.py (1582 LOC) — visits list, mark-paid (visit+queue),
  complete, start-visit, records actions
- __init__.py (14 LOC) — re-exports router, imports all submodules

No external imports from registrar_wizard.py — backward compat
maintained via __init__.py re-exporting router.

Cross-module import added in _visits.py for VisitResponse (defined
in _settings.py).

Tests: app loads OK; frontend 554/554 pass (0 regressions).
